Extract logic for LSP CompletionList creation
